Added corrections for module3-task1
