+++ src/dom-util.js
@@ -1,0 +1,3 @@
+export const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
