+++ src/main.js
@@ -8,6 +8,7 @@ import {createTripDayTemplate} from "./view/trip-day.js";
 import {createTripEventTemplate} from "./view/trip-event.js";
 import {createTripEventEditTemplate} from "./view/trip-event-edit.js";
 import {generateTripEvent} from "./mock/trip-event.js";
+import {render} from "./dom-util.js";
 
 const TRIP_EVENT_COUNT = 15;
 
@@ -16,10 +17,6 @@ const menuHeaderNode = headerNode.querySelectorAll(`.trip-controls h2`)[0];
 const filtersHeaderNode = headerNode.querySelectorAll(`.trip-controls h2`)[1];
 const bodyContainerNode = document.querySelector(`.trip-events`);
 const sortHeaderNode = bodyContainerNode.querySelector(`.trip-events h2`);
-
-const render = (container, template, place) => {
-  container.insertAdjacentHTML(place, template);
-};
 
 const getTripEventsByDays = (tripPoints) => {
   const tripDays = new Map();
@@ -41,7 +38,6 @@ const tripEvents = new Array(TRIP_EVENT_COUNT)
   .fill()
   .map(generateTripEvent)
   .sort((a, b) => a.timeStart - b.timeStart);
-
 
 const tripDays = getTripEventsByDays(tripEvents.slice(1));
