+++ src/view/trip-day.js
@@ -1,15 +1,5 @@
 import {getDateAtShortFormat} from "../util.js";
-
-const REGEX_SYSTEM_DATE = /(\d{2}).(\d{2}).(\d{4})/;
-
-const getSystemFormattedDate = (date) => {
-  const dateArgs = REGEX_SYSTEM_DATE.exec(new Date(date).toLocaleString(
-      `en-US`,
-      {year: `numeric`, month: `2-digit`, day: `2-digit`}
-  ));
-
-  return `${dateArgs[3]}-${dateArgs[2]}-${dateArgs[1]}`;
-};
+import {getSystemFormattedDate} from "../date-util";
 
 export const createTripDayTemplate = (date, index) => {
   return (
